recipetools.create_buildsys_python: initial handling of extras_require
extras_require are optional dependencies. As far as I can tell, these are
optional at runtime. Currently I'm emitting PACKAGECONFIGs which add the
rdeps. I thought about just adding them all to rrecommends, but that'll lead
to a package with bloated deps and no ability for the user to control it.
Ideally, I'd be able to add to RRECOMMENDS from PACKAGECONFIG, I think. We may
want to improve/change how this is implemented going forward.

Signed-off-by: Christopher Larson <kergoth@gmail.com>