bitbake: bb.fetch: simplify mirror tarball handling in try_mirror_url
Rather than checking mirrortarball explicitly, simply check and see if the
localpath basenames match up.

We can't just compare methods, as in the case where an http file is found via
a file mirror, that causes the http fetcher to be run to process the
nonexistent mirror tarball, resulting in an attempt to download from upstream
when we already have the file.

This will make it easier to add the ability to support multiple mirror
tarballs, to implement shallow git mirror tarball support.

Signed-off-by: Christopher Larson <chris_larson@mentor.com>