bitbake.conf: set PATCHRESOLVE to noop by default
Automatically spawning a devshell, rather than doing so on demand, is
questionable default behavior, and is potentially problematic in headless
builds. Further, there are problems with the patch resolver today. Default to
noop, and the user can always opt-in to use of the patch resolver by setting
PATCHRESOLVE to user.

Signed-off-by: Christopher Larson <chris_larson@mentor.com>